requirements.txt
# Byte Riders – MAHE Mobility Challenge 2026
# Track 1: Intent & Trajectory Prediction
# Python >= 3.8

torch>=1.13.0
numpy>=1.21.0
matplotlib>=3.5.0
nuscenes-devkit>=1.1.9
